requirements.txt
pandas 
numpy
streamlit-aggrid